Remove ScenarioContent protocol
